pyproject.toml
[project]
name = "progress_service"
version = "0.1.0"
description = ""
authors = [
    {name = "wszeborj",email = "wszeborj@o2.pl"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "fastapi[standard] (>=0.118.2,<0.119.0)",
    "uvicorn[standard] (>=0.37.0,<0.38.0)",
    "pydantic (>=2.11.9,<3.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "authlib (>=1.6.5,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "itsdangerous (>=2.2.0,<3.0.0)",
    "mypy (>=1.18.2,<2.0.0)",
    "alembic (>=1.17.0,<2.0.0)",
    "pre-commit (>=4.3.0,<5.0.0)",
    "strawberry-graphql[fastapi] (>=0.287.0,<0.288.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.mypy]
plugins = ['pydantic.mypy']
python_version = "3.13"
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
warn_return_any = true
ignore_missing_imports = true

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true
